Refactor listeners to be package private
